fix map sorting issue
